agentic-ai/requirements.txt
# =============================================================================
# AWS Infrastructure Agent — Multi-Module Dependencies
# =============================================================================

# ── Module 1: AWS Strands Agent Framework ────────────────────────────────────
strands-agents>=1.0.0             # AWS Strands Agents SDK (open source)
strands-agents-tools>=0.1.0       # Pre-built tools (current_time, etc.)

# ── Module 2: LangChain Agent Framework ──────────────────────────────────────
langchain>=0.1.0                  # LangChain core framework
langchain-aws>=0.1.0              # AWS integrations (ChatBedrock, BedrockEmbeddings)
langchain-community>=0.0.20       # Community integrations
langgraph>=0.0.20                 # State machine workflows
langsmith>=0.0.80                 # Tracing and evaluation platform

# ── Module 3: CDK Generation & Evaluation ─────────────────────────────────────
aws-cdk-lib>=2.100.0              # AWS CDK library
constructs>=10.0.0                # CDK constructs
cfn-lint>=0.83.0                  # CloudFormation linting (optional)

# ── AWS AgentCore Runtime (deployment platform) ───────────────────────────────
bedrock-agentcore>=0.1.0                  # BedrockAgentCoreApp wrapper
bedrock-agentcore-starter-toolkit>=0.1.0  # agentcore CLI (configure / launch / invoke)

# ── AWS SDK ───────────────────────────────────────────────────────────────────
boto3>=1.38.0
botocore>=1.38.0

# ── Rich console output (demo formatting) ────────────────────────────────────
rich>=13.7.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0              # Load .env for local dev
gitpython>=3.1.40                 # Git repository operations (Module 2)

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.0.0

# ── ISV Partner Integrations (Module 3) ───────────────────────────────────────
# Note: These are optional. The code will use mock mode if API keys are not provided.
# Install these packages and set environment variables to enable real integrations:
#   - PATRONUS_API_KEY for Patronus AI
#   - DEEPCHECKS_API_KEY for Deepchecks (note: may not have free tier)
#   - COMET_API_KEY for Comet ML
patronus>=0.1.0                 # Patronus AI evaluation
deepchecks>=0.18.0              # Deepchecks LLM testing
comet-ml>=3.35.0                # Comet ML experiment tracking
